requirements.txt
streamlit
langchain==0.1.16
groq
crewai==0.28.8
crewai[tools]==0.28.8
openai
langchain_groq==0.1.3
python-dotenv>=1.0.0